requirements.txt
pandas
numpy
simpy
psutil
pendulum
pymoo
Django>=3.0,<4.0
psycopg2>=2.8
datetime
djangorestframework
django-cors-headers
django-import-export